+++ skillfarm/tasks.py
@@ -52,11 +52,14 @@ _update_skillfarm_params = {
 
 @shared_task(**TASK_DEFAULTS_ONCE)
 @when_esi_is_available
-def update_all_skillfarm(runs: int = 0):
+def update_all_skillfarm(runs: int = 0, force_refresh=False):
+    """Update all skillfarm characters."""
     SkillFarmAudit.objects.disable_characters_with_no_owner()
     characters = SkillFarmAudit.objects.select_related("character").filter(active=True)
     for character in characters:
-        update_character.apply_async(args=[character.pk])
+        update_character.apply_async(
+            args=[character.pk], kwargs={"force_refresh": force_refresh}
+        )
         runs = runs + 1
     logger.info("Queued %s Skillfarm Updates", runs)
 
@@ -266,3 +269,28 @@ def update_all_prices():
         return
 
     logger.info("Skillfarm Prices updated")
+
+
+@shared_task(**TASK_DEFAULTS_ONCE)
+def clear_all_etags():
+    logger.debug("Clearing all etags")
+    try:
+        # Third Party
+        # pylint: disable=import-outside-toplevel
+        from django_redis import get_redis_connection
+
+        _client = get_redis_connection("default")
+    except (NotImplementedError, ModuleNotFoundError):
+        # Django
+        # pylint: disable=import-outside-toplevel
+        from django.core.cache import caches
+
+        default_cache = caches["default"]
+        _client = default_cache.get_master_client()
+    keys = _client.keys(":?:skillfarm-*")
+    logger.info("Deleting %s etag keys", len(keys))
+    if keys:
+        deleted = _client.delete(*keys)
+        logger.info("Deleted %s etag keys", deleted)
+    else:
+        logger.info("No etag keys to delete")